refactor(e2e): collapse to single Debug simulation, 1 user per feature
Replace PicatinnyCoverage (parallel scenarios, load profiles) with Debug:
one scenario, atOnceUsers(1), exercises every picatinny feature in sequence.

Scala Debug covers: all picatinny feeders + JWT + transactions
  (PicatinnyScenario), all faker feeders validated via HTTP echo
  (FeederValidationFeeders/Cases), CurrentDateFeeder + JWT round-trip
  (HttpIntegrationFeeders/Cases). One WireMock stubs /echo and /echo/*.

Java/Kotlin Debug covers: all picatinny feeders + JWT + transactions
  (PicatinnyScenario), CurrentDateFeeder + JWT round-trip (HttpIntegration*).
  One WireMock stubs /echo/*.
