+++ examples/java-maven-example/src/test/java/org/galaxio/performance/picatinny/Debug.java
@@ -1,0 +1,43 @@
+package org.galaxio.performance.picatinny;
+
+import com.github.tomakehurst.wiremock.WireMockServer;
+import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
+import org.galaxio.gatling.javaapi.SimulationWithTransactions;
+import org.galaxio.performance.picatinny.cases.HttpIntegrationCases;
+import org.galaxio.performance.picatinny.feeders.HttpIntegrationFeeders;
+import org.galaxio.performance.picatinny.scenarios.PicatinnyScenario;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.*;
+import static io.gatling.javaapi.core.CoreDsl.*;
+import static io.gatling.javaapi.http.HttpDsl.*;
+
+/** Layer-4 e2e debug gate: 1 user, 1 call per picatinny Java feature, no load.
+ *
+ * Exercises: all picatinny feeders + JWT + transactions (via PicatinnyScenario),
+ * then CurrentDateFeeder + JWT round-trip over HTTP (WireMock GET /echo/{ts}).
+ */
+public final class Debug extends SimulationWithTransactions {
+
+    private final WireMockServer mock;
+
+    {
+        mock = new WireMockServer(WireMockConfiguration.options().dynamicPort().globalTemplating(true));
+        mock.start();
+        mock.stubFor(get(urlPathMatching("/echo/.+"))
+            .willReturn(aResponse().withStatus(200).withHeader("Content-Type", "application/json")
+                .withBody("{\"ts\":\"{{request.pathSegments.[1]}}\",\"auth\":\"{{request.headers.Authorization}}\"}")));
+
+        setUp(
+            PicatinnyScenario.apply("Picatinny Debug", "java-debug")
+                .feed(HttpIntegrationFeeders.ts())
+                .exec(HttpIntegrationCases.echo())
+                .injectOpen(atOnceUsers(1))
+                .protocols(http.baseUrl("http://localhost:" + mock.port()))
+        ).assertions(global().failedRequests().count().is(0L));
+    }
+
+    @Override
+    public void after() {
+        mock.stop();
+    }
+}
